finished update task, now need to clean up styling
