+++ db.json
@@ -2,7 +2,7 @@
   "tasks": [
     {
       "id": 1,
-      "important": false,
+      "important": true,
       "title": "CS 185 Misha's Office Hour",
       "day": "2021-03-05T19:00",
       "textInfor": "https://ucsb.zoom.us/j/86623368970"
