+++ src/component/tasks.js
@@ -8,8 +8,7 @@ const Tasks = ({tasks, onDelete, onUpdate}) => {
 
     return (
         <div>
-            {/* {tasks.map((task) => (<Task key={task.id} task={task} onDelete={onDelete} onUpdate={onUpdate}/>))} */}
-            {tasks.map((task) => (<Task key={task.id} task={task} onDelete={onDelete}/>))}
+            {tasks.map((task) => (<Task key={task.id} task={task} onDelete={onDelete} onUpdate={onUpdate}/>))}
         </div>
     )
 }
